WIIS-9442

+++ src/Controller/ReceptionController.php
@@ -408,11 +408,15 @@ class ReceptionController extends AbstractController {
                     foreach ($line->getReceptionReferenceArticles() as $receptionReferenceArticle) {
                         $reference = $receptionReferenceArticle->getReferenceArticle();
                         $refsToUpdate[] = $reference;
+
+                        if (count($line->getReceptionReferenceArticles()) === 0) {
+                            $entityManager->remove($line);
+                        }
+
                         $entityManager->remove($receptionReferenceArticle);
                         $articleRepository->setNullByReception($receptionReferenceArticle);
                     }
                     $reception->removeLine($line);
-                    $entityManager->remove($line);
                 }
 
                 foreach ($reception->getPurchaseRequestLines() as $line) {
@@ -1505,11 +1509,27 @@ class ReceptionController extends AbstractController {
         if($receptionId = json_decode($request->getContent(), true)) {
             $receptionRepository = $entityManager->getRepository(Reception::class);
             $reception = $receptionRepository->find($receptionId);
-            if($reception?->getLines()?->count() === 0) {
+            $getNoLine = $reception?->getLines()?->count() === 0;
+            $recepRefArticles = $reception->getReceptionReferenceArticles();
+            $qte = Stream::from($recepRefArticles)
+                ->map(fn(ReceptionReferenceArticle $recepRefArticle) => $recepRefArticle->getQuantite() != 0)
+                ->filter();
+
+            if ($getNoLine) {
                 $delete = true;
                 $html = "
                     <p>{$translationService->translate('Ordre', 'Réceptions', 'Voulez-vous réellement supprimer cette réception')}</p>
                     <div class='error-msg mt-2'></div>
+                ";
+            }
+            else if (!$getNoLine && ($reception->getStatut()->getNom() === Reception::STATUT_EN_ATTENTE && $qte)) {
+                $delete = true;
+                $html = "
+                <p class='error-msg'>
+                    Cette {$translationService->translate('Ordre', 'Réceptions', 'recéption.')} est en attente de réception, 
+                    la supprimer engendrera la suppression des {$translationService->translate('Ordre', 'Réceptions', 'articles.')} 
+                    {$translationService->translate('Ordre', 'Réceptions', 'Voulez-vous réellement supprimer cette réception.')}<br> 
+                </p>
                 ";
             } else {
                 $delete = false;
